pyproject.toml
[project]
name = ""
version = "0.0.1"
requires-python = "~=3.13"
dependencies = [
    "django",
    "djoser",
    "mysqlclient",
    "psycopg2-binary",
    "pysqlite3",
    "djangorestframework-xml",
    "django-debug-toolbar",
    "python-decouple",
    "djangorestframework-simplejwt",
    "django-axes",
    "bleach",
    "drf-spectacular",
    "django-cors-headers",
    "django-filter",
    "pillow",
    "djangorestframework",
    "django-storages>=1.14.6",
    "boto3>=1.40.6",
    "gunicorn>=23.0.0",
    "uvicorn>=0.35.0",
    "dj-database-url>=3.0.1",
    "whitenoise>=6.9.0",
    "drf-nested-routers>=0.94.2",
]

[dependency-groups]
dev = []

[tool.uv]
package = false

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
